test/requirements.txt
urllib3==1.26.5
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
numpy==1.21.0
Pillow==8.3.1
py-eureka-client==0.9.9
requests==2.25.0
Werkzeug==0.16.1
opencv-python==4.4.0.46
